refactor: Remove useless casting

+++ bst/doalg.c
@@ -18,7 +18,7 @@ int doalg(int n, int k, int Best[]);
 int doalg(int n, int k, int Best[])
 {
     // Represent values by their indices
-    int *elements = malloc(sizeof(int) * (int)n);
+    int *elements = malloc(sizeof(int) * n);
     for (int i = 0; i < n; i++)
         elements[i] = i + 1;
